fix_listに表示する際のものをインナークラスに変更しました
