+++ app/src/main/java/com/example/otkyu/h30fenrir/view/CasarealRecycleViewAdapter.java
@@ -5,6 +5,9 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
+import android.widget.TextView;
 
 import com.example.otkyu.h30fenrir.R;
 import com.example.otkyu.h30fenrir.asynchronous.api.GnaviAPI;
@@ -17,7 +20,7 @@ import java.util.List;
  * Created by YukiOtake on 2018/01/27 027.
  */
 
-public class CasarealRecycleViewAdapter extends RecyclerView.Adapter<CasarealViewHolder> {
+public class CasarealRecycleViewAdapter extends RecyclerView.Adapter<CasarealRecycleViewAdapter.CasarealViewHolder> {
     private List<GnaviResultEntity> list;
     private View.OnClickListener listener;
     private ImgAsyncTaskHttpRequest imgAsyncTaskHttpRequest;
@@ -79,4 +82,21 @@ public class CasarealRecycleViewAdapter extends RecyclerView.Adapter<CasarealVie
     }
 
 
+    class CasarealViewHolder extends RecyclerView.ViewHolder {
+
+        public TextView titleView;
+        public TextView detailView;
+        public LinearLayout linearLayout;
+        public ImageView imageView;
+
+        public CasarealViewHolder(View itemView) {
+            super(itemView);
+            titleView = (TextView) itemView.findViewById(R.id.title);
+            detailView = (TextView) itemView.findViewById(R.id.detail);
+            linearLayout = (LinearLayout) itemView.findViewById(R.id.shop_linearLayout);
+            imageView = (ImageView) itemView.findViewById(R.id.imageView);
+        }
+    }
 }
+
+
